Main result print
The compiled program now prints the result. I was able to do this by writing a program in LLVM IR that will print using (the only function I have found with support without linking) puts. The function is not perfect, especially since it breaks if the output string should be longer than 13 characters, but test results thus far look promising.