Argument Checking
There was a small oversight where a reference would not have its arguments checked (due to references not previously having arguments). This has been resolved.

I was also able to fully implement functions that take void arguments.

There is another problem that I found which needs to be fixed: the beast of precedence. Previously I was able to resolve precedence as an end step of parsing. This cannot be possible with infix functions, since