And, Or, Short circuiting, and lexing
I worked on implementing logical AND and OR. With these improvements came some small updates to the Lexer and Token to accommodate. I also started work on expression levels in the lexer that will track whether the program is in an expression (and thus does not need a new line). However, this is still in progress.