Default Value optimizations, Value rework
I noticed that default values were not being processed correctly in optimization, which led to me changing how Variables operate with Assignments and References. The Variable changes were also helpful in anticipation of adding the formal type system in the near future.

Essentially the change in Variables was to have references and assignments directly tied to their variable instead of merely tied to each other. This results in a tighter binding between parsing and checking packages, but less redundancy and faster processing. Additionally, the greater cohesion will be helpful when working on the type system, since types operate on a Variable-level instead of through the assignment exclusively.