Rudimentary type system -recursion
This is *not* a stable build!
I implemented a simple type system, minus recursive complications. The only possible type is still a number, but this shows the direction that we are going in. Each expression can be typed, and types need to match for operations currently, but not anything else.
We ran into some complications. First of all, not everything is a number. That means that I will need to provide some kind of type annotation system in the future for parameters if for nothing else. Second, the recursion issue, since the function value needs the type of the function, and the function gets its type from its value. Third, when matching parameters, the type system should check type, not merely name and/or position. These are three necessary additions, but three complex issues. Thus, I felt it was wholly appropriate to save here before continuing on.