Simple dynamic dispatch on print
I implemented very simple dynamic dispatch on printing from main for booleans and ints. This is a huge milestone!
Even though the dynamic dispatch is not technically useful now (the type of the return from main can be typed statically), when we add user types and inheritance, static typing specificity will not be available. Thus we needed the dynamic dispatch to work.
Right now I am loading the LLVMIR from the files Any, Bool, and Int. For non-built-in types, the code would be specified in a regular function. Support for this exists, but has not been tested.