Overall Stability improvement
I made some small changes to improve the previous commit. I was able to fix a bug in the if statement that was caused by the translation model reversal. I was able to remove the unused temp.ll file.

More notably, I changed the functionality of the conditional that the types of the branches don't need to match, just intersect. I also made a test case that demonstrates this feature. Furthermore, since printing is done by type, it does not matter that the type checker does not know whether the result is a number or a boolean statically.

A final note: many of the test cases do not pass currently because I am still working on translation for boolean operators. I was previously planning to define them as methods of the Int class, but I may implement translation for them without first, then move on if I still decide to do so. The current issue is how built-in type behavior should be implemented and whether the LLVM extension keywords I invented in the past commit will continue to be supported.