Int and Bool global Operators
As was mentioned in the previous commit, I decided to add support for operators built in first before making them methods of the Int and Bool types. That is what I have done in this commit, with some problems.

I was able to make some changes in the checker and optimizer to allow for bools and ints as operands and results (for example, & takes two bool operands. > takes two int operands but returns a bool).

I was able to get a semblance of a working translation. For some cases, what I have now works, but it is not complete. I offloaded all memory into static memory, which means that if the same function is called recursively, values from the previous frame are overridden. I need to devise some kind of memory management to keep track of what should be allocated and freed in the heap.