Memory and Improved Errors
I was able to transform the problem of bad translation of operators to a problem of memory management. Previous to this commit, the operators would not work if the block they were in was reentrant. This was problematic for any reasonable program because of the ubiquity of recursion.

There are different types of memory management. One safe way is to not free any memory. That is what I have opted to do for the time being.

As programs become more complicated, it becomes more important to have good errors. I created some form of print for each of the different parse structures. Also, I was able to separate compile errors from compiler crash errors for the convenience of the user.

Lastly, I should add that all of the tests currently pass. They run much slower than before because each of the integration tests verifies the result for both the optimized and unoptimized compilation strategies.