Simple and Composite User Types
Groundwork for user defined types, which may have 1 or more fields and may have fields typed of other user defined types.

Behind the scenes the compiler creates constructors and structs for each type. Instantiation of an object is by the type name.

Inheritance, methods, and shadowed types may not yet be supported.