Parsing and set up for Inheritance
I made some parsing additions to accommodate for inheritance. The super list is parsed as a tuple (just as an argument list). The super can contain regular type names, default parent values, or labeled values (where both the type and the default value is given).

Additionally, some work was required on the ..super function, which should select the parent struct necessary to access the requested field. The logic is in place for this to be operational on built-in types. An interesting use of this is documented in the updated language guide.

Currently, the compiler will ignore all inheritance structures parsed.

A couple of other small changes: default field values are allowed and tested for types. Empty types are similarly allowable and tested for. Empty types are most useful when behavior and other fields are inherited from some super type.