Inheritance of Fields
I got inheritance for fields operating properly. There is the dynamic @..super method which will find the super for the needed type number, and the static getMemberOf function that will get the type for which the name is a member.

It should be noted that the constructorDefaultValue test is failing (but only for the optimized compiler setting). This needs to be resolved.